@@ -7,7 +7,7 @@ plugins {
     id 'eclipse'
     id 'maven'
     id 'maven-publish'
-    id "com.jfrog.artifactory" version "4.24.23"
+    id "com.jfrog.artifactory" version "4.29.0"
 }
 
 repositories {
@@ -19,24 +19,25 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-autoconfigure:2.2.2.RELEASE'
-    implementation 'com.auth0:java-jwt:3.10.2'
-    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.10.1'
-    implementation 'org.json:json:20190722'
-    implementation 'org.apache.camel:camel-core:3.4.0'
+    implementation 'org.springframework.boot:spring-boot-autoconfigure:2.7.3'
+    implementation 'jakarta.validation:jakarta.validation-api:3.0.2'
+    implementation 'com.auth0:java-jwt:4.0.0'
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.13.3'
+    implementation 'org.json:json:20220320'
+    implementation 'org.apache.camel:camel-core:3.18.1'
     //Open Source repositories should not link to private repositories
     //implementation 'com.interop.ilp.conditions:interop-ilp-conditions:1.3'
-    testImplementation 'org.springframework:spring-web:5.2.2.RELEASE'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.2.2.RELEASE'
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.2'
-    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.6.2'
-    compileOnly 'org.hibernate.validator:hibernate-validator:6.0.20.Final'
+    testImplementation 'org.springframework:spring-web:5.3.22'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.3'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.9.0'
+    compileOnly 'org.hibernate.validator:hibernate-validator:7.0.5.Final'
 
 }
 
 group = 'org.mifos'
 version = '1.1.0-SNAPSHOT'
-sourceCompatibility = '1.8'
+sourceCompatibility = '17'
 def artifactId = 'ph-ee-connector-common'
 def versionNumber = version
 
