@@ -1,6 +1,6 @@
 package org.mifos.connector.common.channel.dto;
 
-import javax.validation.Constraint;
+import jakarta.validation.Constraint;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
