@@ -10,7 +10,7 @@ package org.mifos.connector.common.mojaloop.dto;
 import java.beans.Transient;
 import java.math.BigDecimal;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 
 import org.hibernate.validator.constraints.Length;
 import org.mifos.connector.common.util.ContextUtil;
