@@ -7,7 +7,7 @@
  */
 package org.mifos.connector.common.mojaloop.dto;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 public class Party {
 
